# swagger.yml
swagger: "2.0"
info:
  description: This is the swagger file that goes with our server code
  version: "1.0.0"
  title: Swagger Rest Article
consumes:
  - application/json
produces:
  - application/json
basePath: /api
...
paths:
...
  /movies/budgetMore/{budget}:
    get:
      operationId: movies.budgetMore
      tags:
        - Movies
      summary: Get movies by more budget
      description: Get movies by more budget
      parameters:
        - name: budget
          in: path
          type: integer
          description: search movies by more budget
          required: true
      responses:
        200:
          description: search movies by more budget
...